% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbqr.gs.R
\name{bentQRgrid}
\alias{bentQRgrid}
\title{Grid search of change points for bent line quantile regression}
\usage{
bentQRgrid(y, x, z, tau)
}
\arguments{
\item{y}{the response}

\item{x}{the covariate that may contain a change point}

\item{z}{the covariates that do not contain a change point}

\item{tau}{quantile level}
}
\value{
bp.est estimates of change points
}
\description{
It finds the change points using grid search. The candidates are equally-spaced grid points from 1\% quantile to 99\% quantile of the covariate that may contain a change point. The change point is found by minimizing the quantile check loss.
}
