% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbqr.gs.R
\name{cbqr.gs}
\alias{cbqr.gs}
\title{Censored bent line quantile regression}
\usage{
cbqr.gs(
  cy,
  tt,
  xx,
  tau,
  delta,
  level = 0.05,
  censoring = "right",
  method = "ks"
)
}
\arguments{
\item{cy}{the response (including the censored response)}

\item{tt}{the covariate that may contain a change point}

\item{xx}{the covariates that do not contain a change point}

\item{tau}{quantile level}

\item{delta}{the censoring indicator (0 if censored)}

\item{level}{significance level. The default is 5\%.}

\item{censoring}{the censoring type. 'right' means censoring from below while ’left’ means censoring from above. The default is ’right’.}

\item{method}{the methods for density estimation involved in asymptotic interval estimation. The options include 'nid', 'ker', and 'ks'. The default is 'ks'.}
}
\value{
beta.est estimates of beta_1, beta_2 along with standard error estimates

t.est estimates of $t$ along with standard error estimates

ci.t confidence interval of change point

ci.beta confidence intervals of beta_1, beta_2

gamma.est estimates of gamma along with standard error estimates
}
\description{
It implements the methodology censored bent line quantile regression Y^* = beta_1 t + beta_2 max(t - t_0, 0) + X gamma + epsilon, subject to censoring. In the first step, it estimates an informative subset. In the second step, it estimates a bent line quantile regression model. The estimation of the change points is based on grid search, and the estimation of regression coefficients is based on rq in the package quantreg. The standard errors and the interval estimates are provided with three options ('nid', 'ker', 'ks'). See details in summary.rq for 'nid' (quotient) and 'ker' (kernel density estimation). 'ks' considers a bias correction using the derived Bahadur representation for the residuals in kernel density estimation. For smaller sample sizes, 'ker' is more recommended as 'ks' may produce wide intervals occasionally.
}
\examples{
m <- 10
n <- 50
tau <- 0.50
cl <- 4
t0 <- 1.5
subj.m <- seq(from = .5, to = 3, length.out = m)
yy <- c()
tt <- c()
cy <- c()
for (j in 1:m) {
 t <- runif(n, 0, 3)
 t <- t[order(t)]
 tt <- c(tt, t)
 y <- subj.m[j] + .5 * t + 2 * pmax(t - t0, 0) + rnorm(n, sd = .1)

 # censoring
 c <- y
 if (min(which(y > cl)) < Inf) {
   c[min(which(y > cl)):length(c)] <- cl
 }

 yy <- c(yy, y)
 cy <- c(cy, c)
}
# covariate
group <- as.factor(sapply(1:m, function(x) rep(x, n)))
xx <- as.matrix(model.matrix(~ group))
xx <- xx[, - 1]
## GAM estimation
delta <- ifelse(yy <= cl, yes = 1, no = 0)
mod.gs <- cbqr.gs(cy = cy, tt = tt, xx = xx, tau = tau, delta = delta)
}
\references{
Li, Chenxi, Ying Wei, Rick Chappell, and Xuming He. "Bent line quantile regression with application to an allometric study of land mammals' speed and mass." Biometrics 67, no. 1 (2011): 242-249.

Tang, Yanlin, Huixia Judy Wang, Xuming He, and Zhongyi Zhu. "An informative subset-based estimator for censored quantile regression." Test 21 (2012): 635-655.
}
